Extract camera upload logic into separate component
